Revert "feature/foldersChats"
