+++ lib/chat/ui/screens/chats.screen.dart
@@ -1,13 +1,18 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
+import 'package:lab_movil_2222/chat/models/chat-participant.model.dart';
 import 'package:lab_movil_2222/chat/models/chat.model.dart';
-import 'package:lab_movil_2222/chat/services/list-general-chats.service.dart';
-import 'package:lab_movil_2222/chat/services/list-group-chats.service.dart';
+import 'package:lab_movil_2222/chat/services/create-personal-chats.service.dart';
+import 'package:lab_movil_2222/chat/services/get-chat.service.dart';
+import 'package:lab_movil_2222/chat/services/list-player-chats.service.dart';
+import 'package:lab_movil_2222/chat/ui/screens/chat-participants.screen.dart';
+import 'package:lab_movil_2222/chat/ui/screens/messages.screen.dart';
 import 'package:lab_movil_2222/chat/ui/widgets/chat-list-item.widget.dart';
 import 'package:lab_movil_2222/chat/ui/widgets/chat-text-description.widget.dart';
-import 'package:lab_movil_2222/chat/ui/widgets/folder-chat.widget.dart';
+import 'package:lab_movil_2222/chat/ui/widgets/participants-list-item.widget.dart';
 import 'package:lab_movil_2222/player/models/player.model.dart';
+import 'package:lab_movil_2222/player/ui/widgets/search-player-input.widget.dart';
 import 'package:lab_movil_2222/shared/widgets/app-loading.widget.dart';
 import 'package:lab_movil_2222/themes/colors.dart';
 import 'package:lab_movil_2222/widgets/header-logos.widget.dart';
@@ -25,35 +30,34 @@ class ChatsScreen extends StatefulWidget {
 }
 
 class _ChatsScreenState extends State<ChatsScreen> {
-  List<ChatModel>? generalChats;
-  List<ChatModel>? groupChats;
+  List<ChatModel>? allChats;
 
   /// list of chats founds
   List<PlayerModel>? foundPlayers;
 
   StreamSubscription? _createChatSub;
-  StreamSubscription? _generalchatsSub;
-  StreamSubscription? _groupchatsSub;
+  StreamSubscription? _chatsSub;
 
-  ListGeneralPlayerChatsService get _listGeneralChatsService =>
-      Provider.of<ListGeneralPlayerChatsService>(context, listen: false);
-  ListGroupPlayerChatsService get _listGroupChatsService =>
-      Provider.of<ListGroupPlayerChatsService>(context, listen: false);
+  CreatePersonalChatService get _createPersonalChatService =>
+      Provider.of<CreatePersonalChatService>(context, listen: false);
+
+  GetChatService get _getChatService =>
+      Provider.of<GetChatService>(context, listen: false);
+
+  ListPlayerChatsService get _listPlayerChatsService =>
+      Provider.of<ListPlayerChatsService>(context, listen: false);
 
   @override
   void initState() {
     super.initState();
-    _generalchatsSub = _listGeneralChatsService.chats$
-        .listen((event) => setState(() => generalChats = event));
-    _groupchatsSub = _listGroupChatsService.chats$
-        .listen((event) => setState(() => groupChats = event));
+    _chatsSub = _listPlayerChatsService.chats$
+        .listen((event) => setState(() => allChats = event));
   }
 
   @override
   void dispose() {
     _createChatSub?.cancel();
-    _generalchatsSub?.cancel();
-    _groupchatsSub?.cancel();
+    _chatsSub?.cancel();
     super.dispose();
   }
 
@@ -85,35 +89,87 @@ class _ChatsScreenState extends State<ChatsScreen> {
             ),
           ),
 
-          /// description text
+          /// search input
           Padding(
             padding: EdgeInsets.fromLTRB(sidePadding, 0, sidePadding, 20),
             child: ChatTextDescription.informationText(
               color: Colors2222.white.withOpacity(0.8),
             ),
           ),
-          if (generalChats == null)
+
+          /// search input
+          Padding(
+            padding: EdgeInsets.fromLTRB(sidePadding, 0, sidePadding, 20),
+            child: SearchPlayersWidget(
+              onValueChange: (text) {
+                setState(() => foundPlayers = text);
+              },
+            ),
+          ),
+
+          /// searching results
+          if (foundPlayers != null) ...[
+            /// list of found chats
+            for (PlayerModel player in foundPlayers!)
+              ParticipantsListItem(
+                context: context,
+                color: Colors2222.white,
+                primaryColor: Colors2222.white,
+                participant: ChatParticipantModel(
+                    displayName: player.displayName, uid: player.uid),
+                createChatCallback: (context) => _createChat(player.uid),
+              ),
+          ]
+
+          /// if no chats loaded show loading icon
+          else if (allChats == null)
             const AppLoading()
 
           /// show a list of all chats
           else ...[
             /// chats items
-            for (ChatModel chat in generalChats!)
+            for (ChatModel chat in allChats!)
               ChatListItem(chat: chat, context: context),
           ],
-          if (groupChats == null)
-            const AppLoading()
-
-          /// show a list of all chats
-          else ...[
-            /// chats items
-            for (ChatModel chat in groupChats!)
-              ChatListItem(chat: chat, context: context),
-          ],
-
-          const ListFolderChat(route: 'Personal'),
         ],
       ),
+    );
+  }
+
+  /// TODO: remove duplicated code
+  void _createChat(String playerId) {
+    print(playerId);
+    if (_createChatSub != null) {
+      return;
+    }
+
+    _createChatSub = _createPersonalChatService.create$(playerId).listen(
+      (response) async {
+        /// validate chat was found
+        if (response.chatId == null) {
+          ScaffoldMessenger.of(context)
+              .showSnackBar(const ChatSnackbar.couldNotCreateChat());
+          return;
+        }
+
+        /// load chat
+        final ChatModel? chat =
+            await _getChatService.getChatWithId(response.chatId!);
+        if (chat == null) {
+          ScaffoldMessenger.of(context)
+              .showSnackBar(const ChatSnackbar.chatNotFound());
+          return;
+        }
+
+        Navigator.pushNamed(context, MessagesScreen.route,
+            arguments: MessagesScreen(chat: chat));
+      },
+
+      /// clean stream
+      onDone: () {
+        _createChatSub?.cancel();
+        _createChatSub = null;
+      },
     );
   }
 }
